chore(typings): drop types.d.ts, move some declaration to other file
